+++ webrtc/src/webrtc_c/api/video/i420_buffer.cc
@@ -28,6 +28,16 @@ WEBRTC_EXPORT struct webrtc_I420Buffer_refcounted* webrtc_I420Buffer_Create(
   auto buf = webrtc::I420Buffer::Create(width, height);
   return reinterpret_cast<struct webrtc_I420Buffer_refcounted*>(buf.release());
 }
+WEBRTC_EXPORT struct webrtc_I420Buffer_refcounted*
+webrtc_I420Buffer_CreateWithStrides(int width,
+                                    int height,
+                                    int stride_y,
+                                    int stride_u,
+                                    int stride_v) {
+  auto buf =
+      webrtc::I420Buffer::Create(width, height, stride_y, stride_u, stride_v);
+  return reinterpret_cast<struct webrtc_I420Buffer_refcounted*>(buf.release());
+}
 WEBRTC_EXPORT int webrtc_I420Buffer_width(
     const struct webrtc_I420Buffer* self) {
   auto buf = reinterpret_cast<const webrtc::I420Buffer*>(self);